Added initial support for ARC validation
Fixes issue #386
